faft: Start devserver precisely
Original we used the start_devserver script to start a devserver. It is a
bash script and forks a separate python process for the devserver. It also
blocks the SIGTERM. So we can't terminate the devserver process actually.

This CL is to fork the devserve python process directly, instead of the bash
script.

An additional check is added to ensure the devserver startup correctly.

BUG=chrome-os-partner:14884,chromium-os:34395
TEST=run_remote_tests.sh --board daisy --remote dut -a \
     "image=/path/to/chromiumos_test_image.bin" TryFwB/control$

Change-Id: I4023e67dd5d09fb40e24f636ccbe1ddecfd32778
Reviewed-on: https://gerrit.chromium.org/gerrit/36424
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>