Revert "Increase power_LoadTest WiFi connection retries"
This reverts commit f17380a078c60f294036d485ae345e66e95887b5.

This improved nothing, so I had another look at it. Seems that the line
"Associating..." we can see in all the logs (even those that fail)
proves that the "Connect" step of the setup went through successfully --
and that is the only one affected by this retry number. Also, it would
add a log message on every failed retry, which I have never seen, so
this usually seems to work on the first try and three retries should
really be enough.

Sorry for missing that in the first place... reverted this and removed
the comment that led me in this direction in the first place (seems to
be no longer valid). Will keep on looking for the problem, but whatever
it is, this was not it.

BUG=chromium-os:35120
TEST=none

Change-Id: I4aacf9bafb441c3dd20cb2039471f22a678e0b69
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/36472
Reviewed-by: Sameer Nanda <snanda@chromium.org>