[autotest] WiFi tests marked skip_test should report N/A.
Currently, a test that goes through site_wifitest that is marked with
skip_test in the controlling dict will only skip running the test code,
and then continue on as normal. This means that autotest sees the test
as finishing correctly, and reports it as a passing test. Since we
didn't run the actual test, we should be reporting back that this was a
not run test. Therefore, we should be raising a TestNAError.

TEST=change all WiFiMatFunc tests to include skip_test, and then do a
run_suite for the wificell suite. Verified that the tests are reported
as not running.
BUG=chromium-os:35709

Change-Id: If5670cf255a99473d351255b27b1fdef8d8f4e2d
Reviewed-on: https://gerrit.chromium.org/gerrit/36534
Tested-by: Alex Miller <milleral@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Ready: Alex Miller <milleral@chromium.org>