faft: Backup the original firmware before re-flashing the firmware
If the arguments 'image' and 'firmware_update' are specified, we will
reimage the OS and reflash the firmware before running a test. Sometime the
reflash may fail and leave the DUT in a broken state. We should backup
the original firmware first. The reset_client method, triggered by reboot
timeout, would restore it to the original firmware.

BUG=chromium-os:23309
TEST=manual, run the following test which reflashes the firmware:
$ run_remote_tests.sh --board link --remote dut -a \
      "image=/path/to/chromiumos_test_image.bin firmware_update=True" \
      TryFwB/control$

Change-Id: Iaf36f6b3b5e79b951cf8650d1d2f0f4937901464
Reviewed-on: https://gerrit.chromium.org/gerrit/36654
Reviewed-by: Vic Yang <victoryang@chromium.org>
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>