Convert platform_Shutdown into a server-side test as it needs a host to reboot.
Originally, platform_Shutdown was written as a client-side job that required
it to be run after a system has rebooted. Instead, platform_Shutdown should
work stand-alone and in order for this to happen, it must be a server-side
test that initiates a reboot on its host object.

BUG=chromium-os:35160
TEST=Ran with cros_run_vm_test platform_Shutdown with 0 state before.

Change-Id: Ia1b0d1642b9b050e928e19b497085bed55b676a6
Reviewed-on: https://gerrit.chromium.org/gerrit/36720
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: Chris Sosa <sosa@chromium.org>
Commit-Ready: Chris Sosa <sosa@chromium.org>