firmware_TouchpadMTB: add robot control for click gesture
Add the robot control for testing taps and clicks. A robot script
name dictionary is defined to map a gesture to its corresponding
robot control script. A number of dictionaries are defined to
map the gesture variations to the corresponding parameters for
the control script.

BUG=chromium-os:35751
TEST=Run the following command in the crosh of a chromebook set up
with the touch robot.
$ DISPLAY=:0 OPTIONS="-m ROBOT" python main.py
It would command the robot to perform a series of gestures including
taps and clicks, and display the test result on the Chrome browser
when all tests are finished.

Change-Id: I017d9d94eca737ec938f6249257ddc5dcea85cc4
Reviewed-on: https://gerrit.chromium.org/gerrit/36893
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>