[autotest] Restrict network3g to a cellular pool.
All devices with modems have been added to |pool:cellular|, and this
suite is now restricted down to this pool. If this is not done, then
when suite_scheduler tries to schedule this suite against a board, it
will see that there are a few trivial tests it can run, and then try to
grab all the machines to run the few trivial tests. This allows for
insanity such as network3g trying to grab 11 machines of a certain board
type, even though no machines with that board have any cellular
components attached to them, and we don't even have 11 machines of that
board in the lab.

TEST=./suite_scheduler.py -t
BUG=None

Change-Id: If1b82f3d2146e74d5047c70e6300e897c32e0d04
Reviewed-on: https://gerrit.chromium.org/gerrit/37060
Tested-by: Alex Miller <milleral@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Alex Miller <milleral@chromium.org>