security_RuntimeExecStack: examine process stack segments
This test examines all the running processes and screams if any of them
have an executable stack. This augments the platform_ToolchainOptions
tests in that it can potentially catch mprotect(PROT_EXEC) mis-use.
Also, it is very fast and could easily live in the bvt so we can catch
exec-stack libraries more quickly.

BUG=None
TEST=passes on link, fails (as expected) on daisy

Change-Id: I19b6cdfda26813c6c2654c913d8bd8a2eef3017d
Signed-off-by: Kees Cook <keescook@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/37079
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>