Do not try write protecting the EC if wp_boot == 1
In case the DUT has the hardware write protection enabled (the WP
screw is installed), the EC write protection action could not be
undone unless the screw is removed.

Do not run the firmware_ECWriteProtect test in case the system is
write protected in HW. Also, remove the ensure_fw_a_boot() method not
used by anyone.

BUG=chrome-os-partner:15835
TEST=manual
  .  run the following test

  ~/trunk/src/scripts/run_remote_tests.sh --board=link --servo \
           --remote=192.168.1.2 firmware_ECWriteProtect/control.dev

  with and without the write protect screw installed. The test passes
  in both cases.

Change-Id: I2e42a7057f2d615456b158d5edd0b3f8a7febca7
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/37123
Reviewed-by: Vic Yang <victoryang@chromium.org>