Fix power_LoadTest stat calculation
This is just a small fix to a bug in my recent power_status.py
refactoring (change I0b55b6) that I noticed once but forgot to actually
upload for some time. When the AbstractStats#refresh method gets called
more than once in incremental mode (which AFAIK is not done in any
current test), it would return erroneous results.

BUG=None
TEST=None

Change-Id: Iee299e8aead4861e375058386befa4ea976a957e
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/38783
Reviewed-by: Sameer Nanda <snanda@chromium.org>