Fix usage of -z tests in site_utils/setup_dev_autotest.sh.
Expressions like [ -z $VAR ] will do the wrong thing if "$VAR" has
spaces in it, and the expression isn't a terribly brilliant idea
when "$VAR" is empty, either.  Ditto for tests with '=' and '!='.
For sanity, fixed the script to quote all test expressions that
might be a problem.

BUG=None
TEST=Run the script, provide a password with a space in it

Change-Id: I6200cbe05e3f1223182f79a52f72856c15c155f5
Reviewed-on: https://gerrit.chromium.org/gerrit/37280
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>