faft: A workaround to wait for client RPC server start-up
We found that the RPC call retrial doesn't work all the time and causes
timeout error happened. It breaks FAFT most of the time. So add this delay
to wait the client RPC server start-up as a work-around. This is a
temporary fix.

BUG=chrome-os-partner:16037
TEST=run a FAFT test and pass.

Change-Id: If1df1f697dc336f2ace63d6d5e5a5393acc32916
Reviewed-on: https://gerrit.chromium.org/gerrit/37431
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>