firmware_TouchpadMTB: Improve the consistency of gaps detection
The NoGapValidator used to create many false negatives in gestures
in which there are no noticeable gaps. The problem is that the
method get_largest_gap_ratio() is not strict enough in filtering
gaps. In this CL, we re-define the largest_gap_ratio as the largest
gap_ratio_with_prev at some point satisfying:

  . curr_gap >= GAP_LOWER_BOUND, and
  . gap_ratio_with_next > RATIO_THRESHOLD_CURR_GAP_TO_NEXT_GAP

  where
  . distance(t) = distance between point[t] and point[t-1]
  . Assume the index of current point is t:
    - prev_gap = distance[t - 1]
    - curr_gap = distance[t]
    - next_gap = distance[t + 1]
  . gap_ratio_with_prev = curr_gap / prev_gap
  . gap_ratio_with_next = curr_gap / next_gap

  In NoGapValidator, the largest_gap_ratio derived above should
be less than the new no_gap_criteria in order to pass the validator
test:

    no_gap_criteria = '<= 1.8, ~ +1.0'

  A number of gestures with and without gaps are added to the
unit tests to ensure that there are no false negatives and
false positives in the test cases.

BUG=chromium-os:35752
TEST=Run the following unittest on a host:
$ cd .../autotest/files/client/site_tests/firmware_TouchpadMTB
$ python tests/mtb_unittest.py
The unit tests should all pass.

Change-Id: I14d1ca36ab37ba059dda1a40e00ab7f746c79002
Reviewed-on: https://gerrit.chromium.org/gerrit/37604
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>