autotest: dns_server: Send more specific exception error
If the backup file goes missing, this might be due to the shill
process restarting while the test runs.  We should raise an
exception that represents the true reason for the test failure.

BUG=chromium-os:35984
TEST=Run desktopui_DocViewing test; remove resolv.conf.bak while
test is running.  Note that the crash error message mentions
this specific condition.

Change-Id: I075e8decb419cbff0b8c3ba9a8d590dd9c26cced
Reviewed-on: https://gerrit.chromium.org/gerrit/37545
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>