pseudomodem: Implement Enable.
This CL adds partial functionality that allows the pseudomodem to be
enabled without registering to a network. The state machine is
implemented as a step process that uses the modem state as its counter.

A minor change to the dbus_std_ifaces.DBusProperties constructor has
also been included, such that user provided properties are handled more
properly. The class also allows a dbus connection to be set at a later
time when this is not possible during initialization.

BUG=chromium-os:34605
TEST=I tested that the enable state changes function properly by sending
dbus messages to the pseudomodem using dbus-send.

Change-Id: I32d57d3664eb4ef3acb0ef46464736b39e8fd955
Reviewed-on: https://gerrit.chromium.org/gerrit/37589
Commit-Ready: Arman Uguray <armansito@chromium.org>
Reviewed-by: Arman Uguray <armansito@chromium.org>
Tested-by: Arman Uguray <armansito@chromium.org>