Autotest: Fix field order in Job in scheduler_models
So it appears that the order of _fields in a model must match the
column ordering in the database. Since max_runtime_mins was just added
it is the last column and must be the last field.

This bug caused parameterized_job_id to be max_runtime_mins, which
would look up the parameterized_job_id of 1440 (24 hours *60 mins),
which used an mario R11 image from the old devserver.

BUG=chromium-os:36183
TEST=Verified scheduler_models.Job returned the incorrect value for
     parameterized_job_id before the fix and the correct value after.

Change-Id: I74407d479cab54f8e8bf3e9e3a668daf4d28d587
Reviewed-on: https://gerrit.chromium.org/gerrit/37760
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Simran Basi <sbasi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>