autotest: Add USB autosuspend whitelist smarts to power_x86Settings.
CrOS only enables USB autosuspend on built-in USB devices.
power_x86Settings presently checks all USB devices for this being
enabled.  This CL adds smarts to evaluate the whitelist file and only
check the built-in USB devices.

It also pork barrels a slight increase in the RAPL PP0 number from 1W
to 1.2W as I was able to get this to fail sporadically when it was
only 1W

BUG=chromium-os:34876
TEST=run test on several x86 platforms

Change-Id: I50c56c0082a0651e9ab02300c537a69510b3f100
Reviewed-on: https://gerrit.chromium.org/gerrit/37797
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Commit-Ready: Todd Broch <tbroch@chromium.org>
Tested-by: Todd Broch <tbroch@chromium.org>