firmware_TouchpadMTB: add the new NoLevelJumpValidator
When drawing with a fat finger or with thumb edge, the 'rainbow'
effect may occur. In this CL, we add the new NoLevelJumpValidator to
detect the effect instead of using a strict linearity validator
which usually incurs false negatives.

The 'rainbow' effect occurs in a line with discontinuous line segments
levels up or down sometimes with different rainbow colors on mtplot
which indicate changing tracking IDs.

When a user draws a horizontal line with thumb edge or with a fat
finger, the line could comprise a horizontal line segment followed by
another horizontal line segment (or just dots) one level up or down,
and then another horizontal line segment again at different
horizontal level, etc. This new NoLevelJumpValidator is implemented
to detect such level jumps effectively. Such level jumps could also
occur when drawing vertical or diagonal lines.

Note that a curvy line without such level jumps would pass this
validator. The reason is that curvy lines are inevitable in gestures
drawn with fat fingers or with thumb edge. We implement this
validator with such allowance. On the other hand, a relaxed
linearity_curvy_criteria is used on LinearityValidator to ensure that
the curve is only allowed to some extent. Hence, a very curvy line
would still fail the test.

A number of unit tests are added to ensure that this new validator
functions properly.

BUG=chromium-os:35752
TEST=Run the following command on a lumpy.
$ DISPLAY=:0 python main.py

Draw lines with thumb edge or with fat fingers in the following
gestures:
 - FAT_FINGER_MOVE_WITH_RESTING_FINGER
 - DRAG_EDGE_THUMB
 - TWO_FAT_FINGERS_TRACKING
Observe that NoLevelJumpValidator fails when the 'rainbow' effect
is observed.

Also draw lines with your finger normally and try not to generate the
'rainbow' effect in the above tests. Observe that NoLevelJumpValidator
should pass even with curvy lines. However, the LinearityValidator
should generate degraded scores when the lines become more curvy.

Run the following unit tests on a host. They should all pass.
$ cd .../firmware_TouchpadMTB
$ python tests/mtb_unittest.py
$ python tests/validators_unittest.py

Change-Id: I6d003a75a2518bcdddd1c4531ff1026496fcbeaa
Reviewed-on: https://gerrit.chromium.org/gerrit/38805
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>