firmware_TouchpadMTB: reduce redundancy in robot_wrapper
In the old robot_script_name_dict, each gesture name is mapped to a
robot control script name. This information is redundant as there
is another method_of_control_command_dict which maps a gesture name
to a get_control_command method.

We will add more gestures for the robot to perform later which will
make the redundancy more severe. In this patch, the
robot_script_name_dict is re-defined to map a get_control_command
method to its related robot control script name so that the
redundancy could be removed.

BUG=chromium-os:35751
TEST=Run the following command on a chromebook with the robot.
$ DISPLAY=:0 OPTIONS="-m ROBOT" python main.py
The tests should run smoothly.

Change-Id: I6b20f03b64757b829bb287bf930403e56f068e67
Reviewed-on: https://gerrit.chromium.org/gerrit/38299
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>