firmware_TouchpadMTB: robot control with finger interaction
In this CL, we add a new robot control mode ROBOT_INT which allows
the robot to perform the following two-finger gestures with finger
interaction:

  - finger_crossing
  - stationary_finger_not_affected_by_2nd_finger_taps
  - resting_finger_plus_2nd_finger_move

Note that in the these new added gestures, the robot needs to pause
per variation so that the user can place and lift a metal finger
repeatedly according to the prompt.

BUG=chromium-os:35751
TEST=Run the following command on a chromebook set up with the robot.
$ DISPLAY=:0 OPTIONS="-m ROBOT_INT" python main.py
The robot should perform both one-finger and two-finger gestures
automatically. In addition, it should also perform some two-finger
gestures with finger interaction.

Change-Id: Ia6527dd0754362e03c3c55a7d379f4ca9b9f19c9
Reviewed-on: https://gerrit.chromium.org/gerrit/38302
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>