firmware_TouchpadMTB: pause at the start of current gesture
It is more natural to display the prompt of adjusting the robot
and to pause at the beginning of a gesture instead of at the end of
the previous gesture. This is also required when such a prompt
and pause is needed at per variation basis rather than per gesture
basis.

BUG=chromium-os:35751
TEST=Run the following command on a chromebook with the robot.
$ DISPLAY=:0 OPTIONS="-m ROBOT" python main.py
Note that the prompt of pressing SPACE between one-finger gestures
and two-finger gestures is displayed at the beginning of
TWO_FINGER_TRACKING gestures. This makes the flow more smoothly.

Change-Id: Ie991a84c96f5fe7d47bc695d208740e0a0d7f57a
Reviewed-on: https://gerrit.chromium.org/gerrit/38300
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>