If other clean up means fail, fall back to a reboot during clean up.
Currently if X crashes or the ui is in a non-running state we will never be
able to successfully run clean up on the machine. This CL adds a reboot
IFF 'restart ui' fails.

Due to the way cros_ui.restart is written this code borrows pieces from it
as opposed to using it. I am deferring the refactor as I expect more than
just this portion of the code to need some rework. crosbug.com/36403.

Fix a bug in cros_ui where CmdError is raised instead of TestError.

TEST=Clean up on a properly working device.
Clean up after stop ui had been executed.
BUG=chromium-os:36107

Change-Id: Ia42664d63022594cf573a07380d47d92936f9bbe
Reviewed-on: https://gerrit.chromium.org/gerrit/38124
Tested-by: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Ready: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>