pseudomodem: Implement Register.
This CL adds partial functionality that allows the pseudomodem to
register with a network. The 3GPP modem uses the provided SIM object to
obtain information on the home network to initiate a fake network scan.

The pseudomodem will now attempt to register to the home network (or
one of the roaming networks, if provided) as part of the Enable cycle.

BUG=chromium-os:34605
TEST=Tested the full enable cycle by sending dbus messages to the
pseudomodem using dbus-send.

Change-Id: I460045251448fc281345918c0836c1659bce00b0
Reviewed-on: https://gerrit.chromium.org/gerrit/38146
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Ready: Arman Uguray <armansito@chromium.org>
Tested-by: Arman Uguray <armansito@chromium.org>