faft: Fix forever waiting for sshing the non-existing 'local' target
The previous CL broke FAFT.
  https://gerrit.chromium.org/gerrit/37944

As the default target_host is 'local'. It makes the test forever waiting
for sshing the non-existing 'local'.

This CL is to change its default value to None. If None, skip the connection
and not do power cycle.

BUG=chrome-os-partner:13970
TEST=Run a FAFT test on Link and passed.

Change-Id: I533cf79c5a3075c410741f5b5c586cf574723b5a
Reviewed-on: https://gerrit.chromium.org/gerrit/38366
Reviewed-by: Gediminas Ramanauskas <gedis@chromium.org>
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>