firmware_TouchpadMTB: send TFK.SAVE key event automatically only in robot mode
This CL fixed a bug about sending the SPACE key event. It should send
the key event to proceed to next test automatically only in the robot
related modes including ROBOT, ROBOT_INT, and ROBOT_SIM modes. While
in the manual mode, it should prompt the user to press keys.

BUG=chromium-os:35751
TEST=Run the following two commands on a chromebook without the robot.
$ DISPLAY=:0 OPTIONS="-s" python main.py
In the manual mode, it should prompt the user to press SPACE to
proceed to next test.

$ DISPLAY=:0 OPTIONS="-m ROBOT_SIM" python main.py
In the ROBOT_SIM mode, it should send the SPACE event to proceed to
next test automatically.

Change-Id: Iefefe282820bc63ffd3260e0f3f01aae03775433
Reviewed-on: https://gerrit.chromium.org/gerrit/38804
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>