Autotest: Log error message when setting power fails.
Currently in site_host if a call to the RPM infrastructure fails, an
exception is raised. However autoserv appears to catch this exception
somewhere and the error message is lost.

Added a logging error message so that the caller has some context as
to why the tests and/or test suites failed.

BUG=chromium-os:36374
TEST=None

Change-Id: I0a21086a6d33adce456bcb512e9a3823e7584058
Reviewed-on: https://gerrit.chromium.org/gerrit/38419
Reviewed-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>
Commit-Ready: Simran Basi <sbasi@chromium.org>