autotest: WiFi: Enable use of ChromeOS device as a router
Break up the "LinuxVMRouter" class into two pieces: A
"LinuxCrosRouter" to represent the general category of
ChromeOS-based routers, and a the "LinuxVMRouter" as a
subclass which is specific to hwsim-based setups.

Expand site_wifitest to go hunting for server executables
in /usr/local.

BUG=chromium-os:25908
TEST=Selected tests in WiFiMatFunc.  Server side iperf has
not been tested yet.  5GHz tests won't work yet because of
regulatory.

Change-Id: I8bc80e17e0f408a13467a64efb2eb15335773dd9
Reviewed-on: https://gerrit.chromium.org/gerrit/38426
Reviewed-by: Tan Gao <tgao@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>