Autotest: Be able to add hosts to AFE when host is down
So if we can't reach a host in the lab and someone wants to add it to
their own AFE, they will run into the issue create_host looping as it
tries to ssh into the device and eventually will fail.

Added a simple ping to check if the host is up before calling create_host.

The previous error caught by the try catch only catches address related
errors, not errors pertaining to unreachable hosts.

BUG=None
TEST=local system

Change-Id: Iaada71afadc642ab205ed011f6cde8f3cd6ed542
Reviewed-on: https://gerrit.chromium.org/gerrit/38464
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Ready: Simran Basi <sbasi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>