firmware_TouchpadMTB: add a new gesture one_finger_to_edge
In the one_finger_tracking gesture, a user is requested to draw a
line. While the line usually looks straight, the linearity test
may fail. This is because there are usually some wobbles when
the finger is entering from the edge or is leaving off the edge.
It may be confusing when a straight line fails the linearity test.

In this CL, we add a new gesture to specificaly test the edge problem.
In this way, we could display the edge problems if any in a clearer
way. The RangeValidator, which checks min/max x/y values, is only
checked on this new gesture.

Except this new edge gesture, other gestures should be performed
without overshooting to the edges. The control parameters in
robot_control module are also adjusted accordingly.

BUG=chromium-os:35752
TEST=Run the following command on a chromebook.
$ DISPLAY=:0 python main.py
Observe the new gesture one_finger_to_edge.

Change-Id: I63018e59bed66ffc8cf4cf00620af53b6205ac71
Reviewed-on: https://gerrit.chromium.org/gerrit/38906
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>