firmware_TouchpadMTB: create default http data dir for test server
We use pyauto to display the http test result file at the end of
the tests. The procedure of starting pyauto tests has been changed
somehow. The pyauto test server checks the existence of the default
http data dir no matter whether it is actually used or not. If it
does not exist, it throws out an error.

In this CL, we create the default data dir if it does not exist.

BUG=chromium-os:35751
TEST=Run the following command on a chromebook with the robot.
$ DISPLAY=:0 OPTIONS="-m ROBOT" python main.py
It should display the test result on Chrome at the end of the tests.

Change-Id: Icfda56186be64575c0d2af7c890ca711f27f7bd8
Reviewed-on: https://gerrit.chromium.org/gerrit/38585
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>