Autotest: Fix adding column max_runtime_mins multiple times.
Currently setup_dev_autotest is broken as we add the column
max_runtime_mins multiple times, in the schema and as a migration.

Reverting the change to the schema so that we add it as a migration
and our dev systems have the columns in the same order as on the
actual servers, also helps with upstreaming in the future.

BUG=None
TEST=Ensure setup_dev_autotest now works properly and the column
     order in the db is correct.

Change-Id: I07a3b4ba0e0ae6a164d71c8db155427eab098d0b
Reviewed-on: https://gerrit.chromium.org/gerrit/38657
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Simran Basi <sbasi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>