firmware_TouchpadMTB: determine the moving direction of a gesture
When computing if there are any reversed motions in a line, we need
to know the moving direction. In this CL, we use the variation
attribute in a gesture to determine the moving direction. This is
more accurate especially when there are only a few points in a
gesture. It will be useful when we would like to compute the
reversed motions in a small segment of a line.

BUG=chromium-os:35752
TEST=Run the following command on a chromebook.
$ cd .../firmware_TouchpadMTB
$ python tests/validators_unittest.py
Those unit tests should all pass.

Change-Id: Ib8a8dceac51b0e121f88b292917319ee927f7596
Reviewed-on: https://gerrit.chromium.org/gerrit/39073
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>