firmware_TouchpadMTB: support segment flags in NoReversedMotionValidator
On gestures such as one_finger_tracking, the NoReversedMotionValidator
usually fails because there may be some reversed motions near both
ends of the line. This is because the capacitive values of the
hardware sensors are not stable when the finger just starts touching
the pad or when the finger is leaving the pad. In order to better
display where the problem is, we support segment flag in
NoReversedMotionValidator:

 - BEGIN: the 10% of points in the begin segment
 - MIDDLE: the 80% of points in the middle segment
 - END: the 10% of points in the end segment
 - BOTH_ENDS: the 10% of points in the begin segment and
              another 10% of points in the end segment.
 - WHOLE: the points in the whole line

BUG=chromium-os:35752
TEST=Run the following command on a chromebook.
$ DISPLAY=:0 python main.py
Observe the test results of NoReversedMotionMiddleValidator and
NoReversedMotionBothEndsValidator.

Change-Id: I9fd953cd896bdfebf36856c67fd0d0b23dbd2388
Reviewed-on: https://gerrit.chromium.org/gerrit/39074
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>