Small fixes in power_Resume refactoring
Change I31c379 broke BVT tests on mario and stumpy. One is related to
over-aggressive error message parsing, the other to an odd firmware
error that has always been present.

This patch tightens error message parsing to ensure only kernel WARNINGs
are treated as fatal, and introduces a whitelist for known firmware
errors that shall not break the test.

BUG=chromium-os:36715,chromium-os:36716
TEST=Submit this and (hopefully) see that the BVTs no longer fail.

Change-Id: I012df7286b0264e01009ca0385fc47308d656922
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/38830
Reviewed-by: Sameer Nanda <snanda@chromium.org>