Handle unexpected TEST_NA status as warnings in the display.
TEST_NA seems to be used by a few tests to indicate that the test
was not attempted due to a missing dependency. Many times there is
a useful explanation in the reason. Therefore, show these as
warnings (yellow) and the reason can be reviewed in the popup.

BUG=chromium-os:36795
TEST=http://truty.i.corp.google.com/results/dashboard/lumpy/lumpy-r25/network.html

Change-Id: Ie8b307d34c0b853ce474ea01e46d7ebfd4aaa4f8
Reviewed-on: https://gerrit.chromium.org/gerrit/38938
Reviewed-by: Mark Koudritsky <kamrik@chromium.org>
Commit-Ready: Mike Truty <truty@chromium.org>
Tested-by: Mike Truty <truty@chromium.org>