power_Resume hotfix for kiev (longer suspend time)
Kiev's tolerance for the RTC interrupt problem seems to be strongly
correlated with the amount of time it is suspended. My tests show that
5 seconds are pretty much guaranteed to fail, 7 seem pretty fine and 6
are somewhere in between.

This patch returns the power_Resume delay to the former 10, which
should hopefully solve our BVT failures for now. In the long run, I
would like to build a database into power_suspend.py that contains
board-specific minimum delays, so that we don't have to hardcode the
lowest common denominator here anymore.

BUG=chromium-os:36886
TEST=Run power_Resume on a kiev.

Change-Id: I7d156e66d1267be597a0cc91c7f4d793421d5dcb
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/39138
Reviewed-by: Derek Basehore <dbasehore@chromium.org>