[autotest] Stop uploading random directories.
Lambdas in python do not capture the values that are passed into them,
they simply capture the scope that they were defined in, and hold onto
that. This means that if they escape that scope, the values they see
when they're run is whatever the environment is at the time that they're
run. Let's walk through the fun example here:

    queue.put(lambda: offload_dir(dir_entry))

Will take essentially a reference to dir_entry. It then places this onto
the queue. We then modify dir_entry by using it as the loop index. Thus,
when we execute the lambda on a seperate thread, it peeks into the scope
of the for loop, pulls whatever the current value of `dir_entry` is, and
uploads that.

So it's time to stop being fancy and do the dumb thing of just pass the
exact directory to upload onto the worker thread instead of throwing
around lambdas.

TEST=offloaded
BUG=chromium-os:36917

Change-Id: Ia0681981c62141ff0127137d6cf028b94ef3e578
Reviewed-on: https://gerrit.chromium.org/gerrit/39145
Commit-Ready: Alex Miller <milleral@chromium.org>
Reviewed-by: Alex Miller <milleral@chromium.org>
Tested-by: Alex Miller <milleral@chromium.org>