cellular: Remove racy verifier assert when disconnecting from call box.
The assert checks to see if the call box is in the disconnecting state
but under some conditions, the modem reports that it is disconnecting
before tearing down its RF connection which means the call box stays in
the connected state.  This assert can safely be ignored because there is
a subsequent check the final disconnect state.

BUG=chromium-os:36974
TEST=cellular_Smoke

Change-Id: I2649cdac10e5d81c53ded74a916da314bb8f56f0
Reviewed-on: https://gerrit.chromium.org/gerrit/39265
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Ready: Thieu Le <thieule@chromium.org>
Tested-by: Thieu Le <thieule@chromium.org>